Revert "Revert "Use platform security domains in keymint/gatekee..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/774c2d84-3c1a-4a2c-9a32-f51eaa436226). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib5cff332798df57222b87b74296ab70f2819e588
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com