Revert "Revert "Use platform security domains in keymint/gatekee..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/caaf8a3e-cd1a-44e1-8261-88eef65055a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iab7801b919ac77661cf80a17dc90277b27798660
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com