Audio CAP: Address ANAPIC comments, Part 3.
Refactor the stable version of parcelables `AudioPolicyForceUse`
and `AudioPolicyForcedConfig` for better structure. The tag of
the `AudioPolicyForceUse` union is equivalent to the old enum
with the same name. The value of the union replaces
the old `AudioPolicyForcedConfig` enum.

The list of allowed values has been moved into the XSD
schema 'audio_policy_engine_configuration.xsd'.

Bug: 364310317
Test: atest VtsHalAudioCoreTargetTest
Change-Id: Ia17e4a9f2d703394862f484b0501a3b3262b135d