Revert "Audio CAP: Address ANAPIC comments, Part 2."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82b4c500-816e-44fb-8fd7-25f82af4e961). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie91fca4ae153447b0d90432fcdfe82f7ee00a13f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com