Revert "Expose more from hwtrust for DICE chain validation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a296549-0c35-4e29-a75d-891ddb743e26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2beedec695d6b0edf566ce62c97ab841258a0d9a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com