Revert "Add an explicitly v3 manifest fragment"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3df9b93e-5a29-49c8-970f-7c44c604fc21). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I996e8222de5ea69606d05d6bfb3376572ad4f9d4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com