Revert "check whether the network interface exists before using it"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/566d67e7-be14-49da-8d87-8c1d55bf385a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379818258
Bug: 379819756
Change-Id: I42967b0aa769466954edecca0fff6cb5aa0af748
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com