Revert "check whether the network interface exists before using it"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e53f0811-9b42-46ee-911f-1d72778bbb57). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379837617
Change-Id: Ia8d1620e426ac2aef5ee4a49e320577998350dc1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com