Revert "audio: Add latest_android_hardware_audio_core_rust"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d0f2299-f997-44a9-87b4-419b67ca3d53). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380086007
Change-Id: I5e2332fee3b0e0022ea95981f9a0855624498fc5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com