Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8bb51ebb-6f5c-4f0f-8757-d8fbb55f419d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380085494
Change-Id: I96a4aeee94ff7d8cdfefb10da449a10a360692ce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com