Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89273d8f-a74f-49d4-8639-02ec72cfda64). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380086148
Change-Id: I3540efdd2a1f3938e35d043e91579abb86b2d986
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com