Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5b64fa99-1c75-4d18-b31b-c30169fd2ed7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380086069
Change-Id: I9197a84ecba9f172998e366b458d251069c70994
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com