Add VTS test that checks for substring "keymint"
This CL exposes a function that allows the checking for a
a substring of the component name in the configuration
descriptor of the leaf certificate for the DICE chain in a
CSR.

The aim is to use this function to check for the substring
"keymint" in the primary KeyMint ("default") instance.

In the bug referenced below, look at third bullet in
comment#3.

Test: m libhwtrust_cxx
      atest libkeymint_remote_prov_support_test
Bug: 376297620
Change-Id: Ic81bf8d074b5afeec011513e7dfbc55a1e5da435