Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/588fd4ff-33c6-45ce-9fef-90c18bd81c19). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380086147
Change-Id: I16facda4833e456ebf67355fdffe9c6d2eaf2d24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com