Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b5e0fecf-5960-4804-a205-b65c4c2234eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380085905
Change-Id: I29f1fd261cc04e29816a6648ac9dc00a2917b5e8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com