Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ab8a126-e904-4eab-8791-15a3c28548fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380084938
Change-Id: I15a4c3ad3c454a6abdd43765ab1b288c863f8165
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com