Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ab92775-3022-49d6-a835-ff8baad0fd0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380085901
Change-Id: I254ef8841e4536ed75071fb8069554db5084b691
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com