Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21918093-35e4-45f9-b3d3-804d36810722). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380157584
Change-Id: Icb106291de127ff673a6e35d3a3ca2b636d1ebfe
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com