Revert "Audio CAP: Address ANAPIC comments, Part 3."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d1634276-779e-4584-8681-09af2356ce77). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380158123
Change-Id: I23955db7e6b69122061ad91cfc340121728b912b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com