Revert "ExternalCameraDevice: increase max bytes per pixel"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2783cb8f-8994-4b5d-841b-741351c2c4a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I92c578cf619d0257a3ea13bd76feb98af98614cd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com