Revert "Define AuthMgr API for client authorization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f0ecb2a-71d2-4ec5-80f9-0aabc9f9435c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380882137
Change-Id: Iaab307dd8d2527a707bfd816ad2a9700d28e2b4f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com