Revert "KeyMint VTS tests for module hash feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3d5e3252-2056-4baf-bbb9-a1a4907cb5a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I418e7981d83e61d2dda12352e926efac3713718d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com