Revert "hwcrypto: Moving hwcrypto files out of staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e34a2277-2192-4ac5-9e98-54458877d5c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 381426084
Change-Id: Iea558f83c27efb7fbd05831f2a543b3b39351dcc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com