Revert "Implement secretkeeper HAL v2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8c18286-cfdf-4234-98b5-5b6d2795c1f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 382376014
Change-Id: I43279b6c2866bb0086a65367b69d759ef01fa1dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com