Revert "Implement secretkeeper HAL v2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7fc9b1de-6773-4f88-98c3-ae90b94e5598). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383543807
Change-Id: I154aa5f6241f3f1f27b48dfbbe474722dd83788b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com