Revert "Update VTS for Bluetooth Socket Offload with RFCOMM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05f2af8c-a864-4448-a73e-d45755309edb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If9c949158392cc55cc1fe1352bd07d3b8212785f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com