Revert "Update default implementation for Bluetooth Socket Offlo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05f2af8c-a864-4448-a73e-d45755309edb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17ca91ed4d08a3c6b84ae4f10e59030e1edbfef6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com