Revert "Add HAL interface for Bluetooth Socket Offload"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05f2af8c-a864-4448-a73e-d45755309edb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b394cb1fc91f6b5fb6924d6caa78b1644f71ed9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com