Revert "Add fuzzer for bluetooth socket AIDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05f2af8c-a864-4448-a73e-d45755309edb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0eb07b6ffb2ecdfc0fcf60165c1b7ab1c7f5de0b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com