Use consistent, correct terminology for vendor API levels in VTS tests.
And make the RKP VTS tests refer to the SDK API levels consistently
(though it's nice to see that in a few places those tests refer to
the correct vendor API levels with the new YYYYMM format). I left
the YYYYMM references as-is since we'll do a proper refactor later
across all VTS tests, but used the __ANDROID_API_<foo>__ constants
everywhere else to be consistent with all the other VTS tests.

There is no official notion of "VSR API level", only "vendor API level".
VSR refers to the Vendor Software Requirement document, so it makes
sense to refer to VSR-15 (for example) since that refers to a specific
version of the requirements, but not to a "VSR-15 device" but rather a
device with vendor API level 35 (in the same way that we don't refer to
CTS-15 devices, but rather to devices with SDK API level 35).

Bug: 384021501
Test: TODO (will do atest VtsAidlKeyMintTargetTest once it's fixed)
Change-Id: I3914c26f81b3bdc886d06cc17f238b5de597ccfc