Revert "Replace channel layout for 13.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/248bd9e5-9329-44ec-93e2-5a171b3296e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384585558
Change-Id: Ibb4b67d5fc284e72716f925900af99714c3ee088
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com