Revert "Replace channel layout for 13.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/80a3585f-33cb-45b0-9591-c996b058bfbd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384584612
Change-Id: I782ca4a6518f8aa745e4043675dfb78deda784e8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com