Revert "Convert SDK API level to vendor API level for Android 15..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/836f24c5-4ab3-4ba2-91da-d0c629920d29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4aceabfd4bb1e0de461d126d09f1d286ed636e50
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com