Revert "Make VtsHalBootAidlTargetTest failures easier to debug."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e14007b0-245e-4a99-937c-19b5bc042c94). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b95096ec7d0cdc26e9bb2b74803455534b73b90
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com