Revert "Convert SDK API level to vendor API level for Android 15..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/156b4c07-69a5-499d-aa13-1012186a3807). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15ab7ee20e6a4f9af655100dc1233778c6305d2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com