Revert "secure_element: Don't enable omapi feature on cuttlefish"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/535274b2-97d7-4f4a-a03d-7a22aa20e7a6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 386116485
Bug: 386116485
Change-Id: Ic3bf472f99c9599c5c898a883ad33c9542079289
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com