Revert "EnvironmentalReverbTest: Modify test parameter values"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b831baf7-30f9-4c1e-bca7-c29155ecfa76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I543591ac026c451b48e7e3e1e3fdaec845739120
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com