Revert "Allow any mode when checking DICE chain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/70375969-8615-48f3-a045-7ea7c825f748). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3b6c50077177a6e6192410fa038d6448dbc60de5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com