Revert "Allow any mode when checking DICE chain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1aad4bb5-deff-4c70-91d5-f8adabf2904c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I632d39ff8160739b306af1f956c473893051f55f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com