Make VTS test for DICE chain more restrictive
This test should check that RKP VM DICE chain conforms to a certain
criterion. That is, the first certificate in a DICE chain that has an
RKP VM marker should have a mode that respects the bootloader's state.
A normal mode should be equivalent to a "locked" bootloader. If the DICE
chain does not contain a certificate with an RKP VM marker, then this
test should be skipped.

Test: none yet
Bug: 388092592
Change-Id: I2717284af456e285477fe9580ef8fe23f564aecc