audio: Update HIDL VTS VtsHalAudioPolicyV1_0TargetTest
Match the change of 'ParameterManagerWrapper' from aosp/3438629.

Do not validate any vendor-side configs with HIDL XSD if it uses
AIDL HAL. This is because the XML config files can still be
present but they are not required to conform to HIDL XSD.

Note that the "hybrid mode" CAP should have a separate test,
and does not need to be covered by this test.

Bug: 379596346
Bug: 456247269
Test: atest VtsHalAudioPolicyV1_0TargetTest
Change-Id: I7d435c6fecc36fd03779aad9c7148c9e9a6c848a
(cherry picked from commit 96828c48268bbf5cd30bb93183b7c8077e4e2bde)