Revert "audio: Fix MonoPipe ownership in StreamRemoteSubmix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c236839-e241-45b6-919a-0986e88419a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If891fd9fac369561f54e9c2143b793fc032ffd16
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com