Revert "audio: Fix MonoPipe ownership in StreamRemoteSubmix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/42c4e376-2f31-4a40-b170-e8838037d8cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7087ed6df874a0f0c30102ee83badce8319b0706
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com