KeyMint: wrapping key must be RSA
The description of the `unwrappingParams` parameter to
`importWrappedKey` incorrectly implies that the wrapping key might be an
AES key.  This is not supported; it must be an RSA key as indicated 20
lines further up.

DO NOT SUBMIT: this is just a test of presubmits on this branch with only a comment change involve.

Test: None, comment change
Change-Id: Iabe6125ba8de0c6273f26fef2ed57e15e67a7311
(cherry picked from commit 615191222e54b8a9601dafd8f3efee3be28415ac)