Revert "KeyMint: wrapping key must be RSA"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e3020853-cd37-49af-b1f8-11fa70bed771). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 397453497
Change-Id: Ib9398f5b6ca429060bf3076dab2701aafe9520fb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com