Revert "KeyMint: wrapping key must be RSA"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48e99653-f142-4485-bcd6-736f93edead2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 397445008
Change-Id: I1eac398705657c724933a82717d838a0b607cb2c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com