Revert "Remove dangling comment"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/229a5cf7-23bf-40a4-85e3-7420c0efdb22). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 398918267
Change-Id: Iff053e5237c700ef718996f613be65d13e8f14b7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com