Revert "[owners] Remove egranata@google.com from automotive/vehi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b93d390b-91bc-4b1d-8b7f-7aeadaef295b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e50287624aabe7f21e35b8938de0a39f7252cc5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com