Revert "secure_element: Don't enable omapi feature on cuttlefish"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4338c366-2790-4cb2-8827-512c298b1c7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 399973797
Change-Id: Ia2981558056a7728d4a5cd537864c5a158fdbb14
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com