Revert "DynamicsProcessing: Add test for MBC band config's prega..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/502ee89d-e85b-4539-b2b4-ff5f26cd633b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I019a7d60189dede4ae6d327dd4e408463560edd0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com