Revert "hwcryptohal: Adding hwcrypto key tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b593b3c9-03b6-4b44-a002-0a912e6a07b8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 402293327
Change-Id: I7b8a12798f53a5ccf2d006152641efdedf68f69a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com