Revert "hwcryptohal: VTS caiman bug"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1b5758a-f354-4551-9931-77ff3e2d9a62). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icbf7e31dd20624800f09289060f6a0cef626c1c3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com