Revert "Fix VtsHalAudioCoreTargetTest test case error"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd9869ff-bd6d-4ed1-988c-753b22e4ba17). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2ea87dc6694f76e13e065238ddc40852137bcfd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com