Revert "Only test data injection for vec3 sensors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6f72894-5f4a-47ad-83c0-e473290bc607). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b6394e23ecb41dd7c1422e11fa7245713d6843c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com