check health HAL version to decide to call getHingeInfo or not
If the corresponding aidl_interface module has frozen set to be false
and the system is built with unfrozen AIDL not allowed to be used, the
HAL acts like the last frozen version, new methods defined in the new
version of AIDL files cannot be used.

In this commit, check the health HAL impl version to decide whether the
getHingeInfo method can be called.

TEST:
  vts case: VtsHalHealthTargetTest
  module: Health/HealthAidl#getHingeInfo/0_android_hardware_health_IHealth_default

Change-Id: I941e5258aa44cc087b14e917c89180d56c6d0134
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>