Revert "hwcryptohal: Adding hwcrypto key tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9b64b34-52d3-43a7-8a6a-3163e08994c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id9f0259fa59210a7c4bbb61ef153b8944053e0db
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com