Revert "Require non-empty names in KeyMintHardwareInfo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c92b820a-0934-45cf-b15a-b1fb0bb127f7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb609510a13ceed3bec77ba82b7e320b38cbe03e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com