Revert "Require non-empty names in KeyMintHardwareInfo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73e992d6-cf23-4e17-beab-b66e74a34d65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4bdf62cf2f4647a8fb686de6bef76a75445c37fe
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com