DO NOT MERGE: Size of vector sidExts should be changed to int64_t
Since the size of vector is int, sidExts is determined by a 32-bit variable, but sidExts requires a size of 64 bits to determine "a000000001", and it fails due to a missing digit.
Since the vector size has been fixed in the int64_t in the main branch, the same modification needs to be applied.

Bug: 407442029
Test: VtsHalBroadcastradioAidlTargetTest PerInstance/BroadcastRadioHalTest#DabTune/1_android_hardware_broadcastradio_IBroadcastRadio_dab
Change-Id: I5465c5e32e1fab38020adb6acc8780cfc98b83bb