Revert "hwcryptohal: Adding hwcrypto key tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a6937a0-59b7-40c5-b32b-436569875c5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I47fd01979dceca7418e123542783d99a464b539f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com