Update BroadcastRadio VINTF fragment to version 3
Updated the AIDL HAL manifest for BroadcastRadio to reflect version 3,
replacing the previous version 2. This change ensures compatibility
with the latest interface definitions and aligns with the current
vendor interface expectations.

This update is necessary to pass VTS tests such as
`vts_treble_vintf_vendor_test`, which validate HAL versioning and
interface compliance.

Test: vts_treble_vintf_vendor_test

Change-Id: I796295e5f8c7b00ecd7cda45500cee384e3127cd