Add jmattis@ to OWNERS for Minimal Telephony
To streamline the integration of Minimal Telephony features
(b/449213597) onto the automotive internal Android 15 branch, jmattis@
is being added to the OWNERS files.

This ownership will facilitate the merging of necessary changes more
easily. We anticipate that there will be a lot of changes for this
particular repo that need to be merged in. This OWNERSHIP change will be
reverted once the set of Minimal Telephony changes are successfully
landed.

Bug: 449213597
Test: N/A (OWNERS file change only)
Change-Id: I2d312cbc7c7a0804e40ecd74fb6f8132f4b67d75